Completing code
Now all the code should be done, excepting minor mistakes

+++ orderprocessingplatform/src/main/kotlin/pt/unl/fct/iadi/orderprocessingplatform/pricing/PriceCalculator.kt
@@ -3,5 +3,5 @@ package pt.unl.fct.iadi.orderprocessingplatform.pricing
 import pt.unl.fct.iadi.orderprocessingplatform.domain.Order
 
 interface PriceCalculator {
-    fun calculateTotalPrice(order: Order): Float
+    fun calculateTotalPrice(order: Order): Double
 }
